Add translations on login
